Post Auto-Calculation cleanup
Created component package within editor;
focused functionality of CharmPanel;
created Notches class to simplify managing equipped & max notches;
isolated autoCalc ActionListener in AutoCalcActivator;
severely cleaned and labelled SaveEditor;
relocated charmNames into CharmPanel;
renamed HKSM\panel to HKSM\app\panel
