+++ HKSM/app/editor/SaveEditor.java
@@ -6,8 +6,6 @@ import java.util.List;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.io.File;
-import java.util.ArrayList;
-import java.util.Collections;
 
 import javax.swing.AbstractButton;
 import javax.swing.JCheckBox;
@@ -23,14 +21,15 @@ import javax.swing.ToolTipManager;
 
 import com.google.gson.JsonObject;
 
-import HKSM.app.editor.Listeners.IntField;
+import HKSM.app.editor.component.CharmPanel;
+import HKSM.app.editor.component.Notches;
 import HKSM.data.SaveLoader;
 import HKSM.app.editor.Listeners.DocChecker;
+import HKSM.app.editor.Listeners.AutoCalcActivator;
 
 /**
  * 
  * @author Kristian Thorpe <sfgmugen@gmail.com>
- *
  */
 @SuppressWarnings("serial")
 public class SaveEditor extends JFrame {
@@ -38,54 +37,11 @@ public class SaveEditor extends JFrame {
 	public JsonObject json;
 	public File fObject;
 	
-	//CHARM_NAME_23_BRK   Fragile Heart (Repair)
-	//CHARM_NAME_24_BRK   Fragile Greed (Repair)
-	//CHARM_NAME_25_BRK   Fragile Strength (Repair)
-	
-	//CHARM_NAME_36_A White Fragment
-	//CHARM_NAME_36_B Kingsoul
-	//CHARM_NAME_36_C Void Heart
-	
-	private static String[] charmNames = new String[]{
-			"Gathering Swarm",
-			"Wayward Compass",
-			"Grubsong",
-			"Stalwart Shell",
-			"Baldur Shell",
-			"Fury of the Fallen",
-			"Quick Focus",
-			"Lifeblood Heart",
-			"Lifeblood Core",
-			"Defender's Crest",
-			"Flukenest",
-			"Thorns of Agony",
-			"Mark of Pride",
-			"Steady Body",
-			"Heavy Blow",
-			"Sharp Shadow",
-			"Spore Shroom",
-			"Longnail",
-			"Shaman Stone",
-			"Soul Catcher",
-			"Soul Eater",
-			"Glowing Womb",
-			"Fragile Heart",
-			"Fragile Greed",
-			"Fragile Strength",
-			"Nailmaster's Glory",
-			"Joni's Blessing",
-			"Shape of Unn",
-			"Hiveblood",
-			"Dream Wielder",
-			"Dashmaster",
-			"Quick Slash",
-			"Spell Twister",
-			"Deep Focus",
-			"Grubberfly's Elegy",
-			"Void Soul"
-	};
-	
 	public SaveEditor(String path, String title){
+		
+		/* ****************** *
+		 * PRE-INITIALIZATION *
+		 * ****************** */
 		ToolTipManager.sharedInstance().setInitialDelay(0);
 		ToolTipManager.sharedInstance().setDismissDelay(20000);
 		String p = "playerData";
@@ -99,10 +55,13 @@ public class SaveEditor extends JFrame {
 		
 		JsonObject playerData = json.getAsJsonObject(p);
 		
+		/* ******************** *
+		 * FRAME INITIALIZATION *
+		 * ******************** */
 		this.setTitle(title);
 		this.setLayout(new BorderLayout());
 		
-		JTabbedPane tab = new JTabbedPane();
+		JTabbedPane tabs = new JTabbedPane();
 		JMenuBar menu = new JMenuBar();
 		
 		JMenu file = new JMenu("File");
@@ -116,14 +75,13 @@ public class SaveEditor extends JFrame {
 				SaveLoader.saveSave(fObject, json);
 			}
 		});
-			
-		JPanel inventoryEditor = new JPanel();
-//		JScrollPane inventoryScroll = new JScrollPane(); // not yet implemented
-		tab.addTab("Inventory", inventoryEditor);
 		
+		/* ************* *
+		 * INVENTORY TAB *
+		 * ************* */
+		JPanel inventoryEditor = new JPanel();
 		inventoryEditor.setLayout(new GridLayout(0,3));
 		
-		// This all could probably be shortened into a for loop of some kind
 		inventoryEditor.add(new SaveField(json, new String[]{p, "maxHealthBase"}, "HP", SaveField.INTEGER));
 		inventoryEditor.add(new SaveField(json, new String[]{p, "MPReserveMax"}, "MP", SaveField.INTEGER));
 		inventoryEditor.add(new SaveField(json, new String[]{p, "nailDamage"}, "Nail Damage", SaveField.INTEGER));
@@ -148,34 +106,23 @@ public class SaveEditor extends JFrame {
 		inventoryEditor.add(new SaveField(json, new String[]{p, "hasDreamNail", "hasDreamGate", "dreamNailUpgraded"}, "Dreamnail", SaveField.DREAMNAIL));
 		inventoryEditor.add(new SaveField(json, new String[]{p, "hasAcidArmour"}, "Ismas", SaveField.BOOL));
 		
-		//Health	| Mana 		| Nail
-		//Geo		| Notches   | Steel Soul
-		//---------------------------------
-		//N-Art1 	| N-Art2	| N-Art3
-		//Spell1	| Spell2	| Spell3
-		//---------------------------------
-		//Dash		| Walljump	| SuperDash
-		//2xJump	| Dreamnail | Ismas
-		//---------------------------------
-		//Lantern	| tram pass | quill
-		//City key  | sly key   | white key
-		//ww key    | spa key   | kings brand
+		tabs.addTab("Inventory", inventoryEditor);
 		
-		JPanel charmTab = new JPanel();
-		charmTab.setLayout(new BorderLayout());
+		/* ********* *
+		 * CHARM TAB *
+		 * ********* */
 		
+		// Notches
 		int usedNotches = playerData.get("charmSlotsFilled").getAsInt();
 		int totalNotches = playerData.get("charmSlots").getAsInt();
+		Notches notches = new Notches(playerData, usedNotches, totalNotches);
 		
-		IntField eNotch = new IntField(usedNotches);
-		IntField mNotch = new IntField(totalNotches);
+		notches.equipped().getDocument().addDocumentListener(new DocChecker(playerData, "charmSlotsFilled", notches.equipped()));
+		notches.max().getDocument().addDocumentListener(new DocChecker(playerData, "charmSlotsFilled", notches.max()));
 		
-		// Now uses boxed listeners
-		eNotch.getDocument().addDocumentListener(new DocChecker(playerData, "charmSlotsFilled", eNotch));
-		mNotch.getDocument().addDocumentListener(new DocChecker(playerData, "charmSlots", mNotch));
-		
-		JCheckBox oc = new JCheckBox("Overcharmed");
-		oc.addActionListener( new ActionListener() {
+		// Overcharmed flag
+		JCheckBox overcharmed = new JCheckBox("Overcharmed");
+		overcharmed.addActionListener( new ActionListener() {
 		      public void actionPerformed(ActionEvent actionEvent) {
 		          AbstractButton abstractButton = (AbstractButton) actionEvent.getSource();
 		          boolean selected = abstractButton.getModel().isSelected();
@@ -183,59 +130,49 @@ public class SaveEditor extends JFrame {
 		        }
 		 });
 		
+		// Auto-Calculator
+		JCheckBox autoCalc = new JCheckBox("Auto-Calculate");
+		autoCalc.addActionListener(new AutoCalcActivator(playerData, notches, autoCalc, overcharmed));
+		
+		// Upper charm panel
 		JPanel charmTop = new JPanel();
 		charmTop.setLayout(new GridLayout(0,2));
 		charmTop.add( new JLabel("Equipped Notches") );
 		charmTop.add( new JLabel("Max Notches") );
-		charmTop.add( eNotch );
-		charmTop.add( mNotch );
+		charmTop.add( notches.equipped() );
+		charmTop.add( notches.max() );
 		
-		JCheckBox autoCalc = new JCheckBox("Auto-Calculate");
-		autoCalc.addActionListener(new ActionListener(){
-			@Override
-			public void actionPerformed(ActionEvent ae){
-				if(autoCalc.isSelected()){
-					int tNotches = 0;
-					for( int i = 0; i < 36; i++){
-						String s = Integer.toString(i+1);
-						boolean eq = playerData.get("equippedCharm_" + s).getAsBoolean();
-						int co = playerData.get("charmCost_" + s).getAsInt();
-						if( eq )
-							tNotches += co;
-					}
-					playerData.addProperty("charmSlotsFilled", tNotches);
-					eNotch.setText(Integer.toString(tNotches));
-					oc.setSelected(eNotch.greaterThan(mNotch));
-				}
-			}
-		});
 		charmTop.add( autoCalc );
-		charmTop.add( oc );
+		charmTop.add( overcharmed );
 		
+		// Lower charm panel
 		JPanel charmBottom = new JPanel();
 		charmBottom.setLayout(new GridLayout(0,2));
-		List<CharmPanel> charmList = new ArrayList<CharmPanel>();
-		for( int i = 0; i < 36; i++){
-			CharmPanel charm = new CharmPanel(i, charmNames[i], playerData, autoCalc, oc, eNotch, mNotch);
-			charmList.add(charm);
+		List<CharmPanel> charmList = CharmPanel.createCharmPanels(playerData, notches, autoCalc, overcharmed);
+		
+		for(CharmPanel each : charmList){
+			charmBottom.add(each);
 		}
-		Collections.sort(charmList);
-		for( int i = 0; i < 36; i++){
-			charmBottom.add(charmList.get(i));
-		}
+		
+		JPanel charmTab = new JPanel();
+		charmTab.setLayout(new BorderLayout());
+		
 		charmTab.add(charmTop, BorderLayout.PAGE_START);
 		charmTab.add(charmBottom, BorderLayout.CENTER);
 		JScrollPane charmEditor = new JScrollPane(charmTab);
-		tab.addTab("Charms", charmEditor);
+		
+		
+		/* ******************* *
+		 * POST-INITIALIZATION *
+		 * ******************* */
+		tabs.addTab("Inventory", inventoryEditor);
+		tabs.addTab("Charms", charmEditor);
 		
 		this.add(menu, BorderLayout.PAGE_START);
-		this.add(tab, BorderLayout.CENTER);
-		
+		this.add(tabs, BorderLayout.CENTER);
 		this.pack();
-		
 		this.setSize(450, 400);
 		this.setLocationRelativeTo(null);
 		this.setVisible(true);
 	}
-	
 }
